Move stream handling to an async method, clean up example
This is a pretty large commit, there are several things I was interested
in achieving

1. Moving the handling of the Sink + Stream objects into an async
   context. It's much less error prone there because we are in the
   "control flow register" [1]
2. Exploring what an API which allows connecting (and disconnecting)
   mutliple streams to the repo dynamically would look like. This would
   fit much more naturally with the way I would expect people to use a
   library like this (e.g. accepting TCP connections from individual
   peers as a stream + sink).
3. Clean up the examples a bit so it's easier to understand what they
   are doing.

To that end the largest change is the `RepoHandle` type and its
`connect_stream` method, which takes a stream and returns a future which
must be driven to process messages for the given stream.

In the example code I've split the example up into multiple files,
implemented the APIs for the relay server and the peer servers
separately, and introduced an abstraction over the stream of messages
from the relay peer (it's obviously still very incomplete as this is
just an example).

[1]: https://without.boats/blog/the-registers-of-rust/