Add a feature flag for tokio specific functionality
To avoid dragging in a tokio dependency for people who aren't using
tokio, and to ensure that the core logic of this library is runtime
independent, introduce `RepoHandle::connect_stream`, which operates over
arbitrary `Stream + Sink` combinations and put the tokio specific
wrapper behind a feature flag. Also add a websocket implementation for
tokio based on `tokio-tungstenite`.

Explain