Use BoxFuture rather than Box<dyn Future<...>> in Storage
The future type returned by `Storage` methods was `Box<dyn Future<....>
+ Unpin` but boxed futures returned by `FutureExt::boxed` are actually
`Pin<Box<Future<...>>>`. This is especially important when dealing
with `async` blocks, which are not `UnPin` and so we _must_ return
`Pin<...>`.

Modify the `Storage` methods to return `futures::BoxFuture<'static,
..>`.

While we're here also make the methods of the `Storage` trait non
optional as it's confusing to suggest that users don't need to implement
them.