Add more descriptive network error
Problem: the `NetworkError` which is returned by the streams and sinks
passed to `Repo.new_remote_repo` is uninformative. All the user knows
when such an error appears in the logs is that there was some kind of
error on the sink/stream.

Solution: Add a `String` field to the `NetworkError` so that it can be
printed in logs.