DRI out the tests
Problem: the integration tests contained a lot of repeated logic for
forwarding messages from one peer to another and for managing tokio
runtimes. This made the tests a little hard to read and a little
fragile.

Solution: replace the repeated routing logic with a simple wrapper
around a pair of `tokio::sync::mpsc::Channel`s and use `tokio::test` so
that we don't need to manage runtimes or synchronise with channels in
the tests.

In the process of implementing this solution I also did a few other
things:

* Fixed bugs in several places where we were failing to `poll_close`
  sinks which we are no longer using
* Used `test_log::test` to run all the tests so we get nice logs from
  `tracing`
* Propagated a tracing span into the thread which repo spawns so that
  spans from within the repo run loop are annotated with the repo ID of
  the running repo.