reply to websocket pings
Context: The websocket protocol contains a ping/pong message type which
is intended for use to check if the other end is still alive.The js
implementation of automerge-repo expects clients to reply to websocket
ping messages promptly and if the client does not reply it will
disconnect them.

Problem: the websocket client included in this library does not respond
to pings, consequently the js server would disconnect the client.

Solution: return a future from `Repo::connect_tungstenite` which the
client must drive. This future responds to websocket pings.