Use base58 encoded UUIDs for document IDs
Problem: the automerge-repo JS implementation uses base68 encoded UUIDs
for it's document IDs. In this library we accept arbitrary strings.

Solution: In preparation for full compatibility with the JS implementation
constrain document IDs to be a UUID. This shouldn't have any
compatibility problems for most users because by default when we create
a document ID we use a UUID.