Clean up tests  a little (#59)
* Add more descriptive network error

Problem: the `NetworkError` which is returned by the streams and sinks
passed to `Repo.new_remote_repo` is uninformative. All the user knows
when such an error appears in the logs is that there was some kind of
error on the sink/stream.

Solution: Add a `String` field to the `NetworkError` so that it can be
printed in logs.

* update `RepoHandle::connect_stream` to split stream and sink

Problem: often the `Stream` and `Sink` you want to use for communication
are separate objects (for example, in the case of using
`tokio::sync::mpsc::Channel` pairs). Forcing the user to combine these
objects into one is awkward, requiring lots of boilerplate.

Solution: accept the `Stream` and `Sink` as separate arguments. This has
a much smaller cost tham forcing users to combine the stream and sink
because most real-life combined streams can be split using
`StreamExt::split`.

* Move tests into single module

Problem: each test in the tests/ subdirectory is compiled as a separate
crate, increasing compile times and making it harder to share common
code between tests.

Solution: move the tests into a single module in `tests/network` where
they are all compiled together.

* DRI out the tests

Problem: the integration tests contained a lot of repeated logic for
forwarding messages from one peer to another and for managing tokio
runtimes. This made the tests a little hard to read and a little
fragile.

Solution: replace the repeated routing logic with a simple wrapper
around a pair of `tokio::sync::mpsc::Channel`s and use `tokio::test` so
that we don't need to manage runtimes or synchronise with channels in
the tests.

In the process of implementing this solution I also did a few other
things:

* Fixed bugs in several places where we were failing to `poll_close`
  sinks which we are no longer using
* Used `test_log::test` to run all the tests so we get nice logs from
  `tracing`
* Propagated a tracing span into the thread which repo spawns so that
  spans from within the repo run loop are annotated with the repo ID of
  the running repo.