forward request messages to connected peers
Context: when a peer requests a document from us we often don't want to
just check our own storage, we also want to ask all other peers we are
connected to if they have the document. This is the case for e.g. a sync
server which is relaying documents between two peers.

Problem: the current immplementation just checks local storage

Solution: when we receive a request, first send a request out to all the
peers we are connected to and once they have all responded only then
send a response to the requestor.