Only announce to new peers when polling storage
Problem: whenever a storage operation wakes us we announce the new
document to all the peers we are connected to. This was intended to
announce documents when loading, but it means we also announce every
time we save.

Solution: only announce to peers whom we do not currently have a
connection when polling storage.