Add `Repo::peer_state` to find out about the sync state w.r.t a peer and doc
Problem: It's often useful to know what the remote state of a document
we are synchronizing is. We have a lot of information about this at the
automerge level but it's not exposed to users of `Repo`.

Solution: Expose automerge sync state information as well as keep track
of when we last received and sent messages.