Load documents when receiving new network events
Problem: when a peer responds to sync messages for a document which it
has not currently loaded it doesn't attempt to load the document from
disk. This means that attempting to request a document whilst
simultanously syncing it with another machine could result in the
dochandle resolving to an empty document.

Solution: ensure that the document is loaded if not already loaded when
the sync message is received.