Add Repo::peer_conn_info_changes to listen to peer connection state
Problem: applications often need to introspect the state of a connection
with a peer in order to debug synchronization issues. For example, if
two users are on a video call and one notices that they have made
changes which the other has not received, they need to determine if the
issue is with the network connection, or a bug in the application, or a
bug in automerge-repo.

Solution: provide Repo::peer_conn_info_changes, which is a stream of
changes to the connection to a particular peer. This includes the last
time we received and sent a message to the peer for any document, and
the last sent/receive for each document, as well as the last acked and
sent heads for each document.

This patch implements this logic very inefficiently, scanning every
document on every change to any document. If the API turns out to be
useful and performance is a problem then we can investigate ways to
optimize this.