Return future from `Repo::new_remote_repo` which resolves on disconnect
Problem: when a remote repo disconnects applications need to a) know
about this and b) decide whether to reconnect. The
`Repo::new_remote_repo` method currently just starts the connection and
then returns, which means the caller has to do their own connection
monitoring by intercepting streams and so on. This is verbose and
creates a lot of duplicated code for each transport.

Solution: return a future from `Repo::new_remote_repo` which resolves
when the connection completes. The output of the future is a
`ConnFinishedReason` which gives some more information about what
happened so that the caller can decide whether to reconnect or not.